+++ packages/kit/test/apps/basics/src/routes/load/layout-props/a.svelte
@@ -1,0 +1,2 @@
+<h2>On a</h2>
+<a href="/load/layout-props/b">to b</a>